started to create the map view in progression page, finished the ability to edit an exsisting item
I decided to let users choose what item they want to edit using a dropdown menu, code was added at the end of both the add item and edit item function to update the dropdown menu with the new item,

editing an item works by: getting the values for the changes and what item was selected, then going through the text file like a list until the line matches the dropdown. then the changes are made.

The map was just added in, still working on working it out